requirements.txt
SpeechRecognition==3.10.0
pydub==0.25.1
pysrt==1.1.2
requests==2.31.0
openai-whisper
